Refactored the error checking system
# IMPORTANT PARADIGM CORRECTION:
In Java, the concept of Exceptions and "errors" are synonymous. However, in Kotlin, Exceptions are seen as unforeseen, unexpected errors which need not be caught.

Hence, canonical usage of `Exception`s should solely be limited to errors which are not expected of a function call.
E.g. NO_INTERNET / NOT_CONNECTED_TO_SCHOOL_WIFI are all not exceptions in the case of the SendATS function because they are in the context where such scenarios are things which are EXPECTED secondary outcomes of their respective erroneous situations.

Hence the refactoring from usage of Exceptions to an explicit `CanErr` interface / custom `sealed class`es (monads) for all SPMobileAPI exposed return types