GGEZ - Managed to figure out how to support vertical scrolling interactions to expand/collapse the CalendarView
A shit ton of components involved in this, I'll prolly not even know how I managed to do it tomorrow, so here goes...

Overview of changes:

`AppBarStateListener` is a wrapper around `AppBarLayout.OnOffsetChangedListener`, which essentially gives a single value, `verticalOffset`, a lot more context, which are listed in the `AppBarStateListener.State` monad.

`ScrollingCalendarBehavior` is set to the default behavior, after realizing that the default behavior is much smarter that just a single boolea constant, since it is intergrated with the `AppBarLayout`

An `OnOffsetChangedListener` is added to the `HomeActivity.onCreate@show_hide_calendar_AppBar` code block, and it implements the specialmemefresh `AppBarStateListener` to generate the cool animations and UX oohs and aahs

The `content_home.xml` layout is drastically simplified after finding out that all the additional wrapping isn't actually necessary. In fact, the abstraction of the layout file is in itself unecessary, but is left there for future-proofing purposes.

The ``